% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19VaccineAesiIncidenceCharacterization.R
\docType{package}
\name{Covid19VaccineAesiIncidenceRate-package}
\alias{Covid19VaccineAesiIncidenceRate}
\alias{Covid19VaccineAesiIncidenceRate-package}
\title{Covid19VaccineAesiIncidenceRate: COVID-19 Vaccine AESI Incidence Characterization}
\description{
COVID-19 Vaccine AESI Incidence Characterization
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ohdsi-studies/Covid19VaccineAesiIncidenceRate}
  \item Report bugs at \url{https://github.com/ohdsi-studies/Covid19VaccineAesiIncidenceRate/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Authors:
\itemize{
  \item Azza Shoaibi
  \item Anthony Sena
  \item Patrick Ryan
}

}
\keyword{internal}
